# services/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateIn": {
        "properties": {
          "key": {
            "maxLength": 64,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Key",
            "type": "string"
          },
          "yaml": {
            "maxLength": 500000,
            "title": "Yaml",
            "type": "string"
          }
        },
        "required": [
          "yaml",
          "key"
        ],
        "title": "CreateIn",
        "type": "object"
      },
...
      "DraftIn": {
        "properties": {
          "yaml": {
            "maxLength": 500000,
            "title": "Yaml",
            "type": "string"
          }
        },
        "required": [
          "yaml"
        ],
        "title": "DraftIn",
        "type": "object"
      },
      "DraftOut": {
        "properties": {
          "issues": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Issues",
            "type": "array"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          },
          "yaml": {
            "title": "Yaml",
            "type": "string"
          }
        },
        "required": [
          "valid",
          "issues",
          "key",
          "yaml"
        ],
        "title": "DraftOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Published": {
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "key",
          "version"
        ],
        "title": "Published",
        "type": "object"
      },
...
      "Validation": {
        "properties": {
          "issues": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Issues",
            "type": "array"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid",
          "issues"
        ],
        "title": "Validation",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "Version": {
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "published_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published By"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "yaml": {
            "title": "Yaml",
            "type": "string"
          }
        },
        "required": [
          "key",
          "version",
          "status",
          "yaml",
          "published_by",
          "published_at"
        ],
        "title": "Version",
        "type": "object"
      },
      "VersionBrief": {
        "properties": {
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "published_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published By"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "published_by",
          "published_at"
        ],
        "title": "VersionBrief",
        "type": "object"
      },
      "WorkflowSummary": {
        "properties": {
          "has_draft": {
            "title": "Has Draft",
            "type": "boolean"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "latest_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Version"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "key",
          "title",
          "latest_version",
          "has_draft"
        ],
        "title": "WorkflowSummary",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Nova API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workflows": {
      "get": {
        "operationId": "list_workflows_api_v1_workflows_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkflowSummary"
                  },
                  "title": "Response List Workflows Api V1 Workflows Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Workflows",
        "tags": [
          "workflows"
        ]
      },
      "post": {
        "operationId": "create_workflow_api_v1_workflows_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/v1/workflows/{key}/draft": {
      "get": {
        "operationId": "get_draft_api_v1_workflows__key__draft_get",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Draft",
        "tags": [
          "workflows"
        ]
      },
      "put": {
        "description": "Saves even when invalid: a draft is work in progress; publish is the gate.",
        "operationId": "put_draft_api_v1_workflows__key__draft_put",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Put Draft",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/v1/workflows/{key}/publish": {
      "post": {
        "operationId": "publish_api_v1_workflows__key__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Published"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Publish",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/v1/workflows/{key}/validate": {
      "post": {
        "description": "Validates the posted YAML, or the saved draft when there's no body.",
        "operationId": "validate_api_v1_workflows__key__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DraftIn"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Validation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Validate",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/v1/workflows/{key}/versions": {
      "get": {
        "operationId": "list_versions_api_v1_workflows__key__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VersionBrief"
                  },
                  "title": "Response List Versions Api V1 Workflows  Key  Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Versions",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/v1/workflows/{key}/versions/{version}": {
      "get": {
        "operationId": "get_version_api_v1_workflows__key__versions__version__get",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Version"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Version",
        "tags": [
          "workflows"
        ]
      }
    }
  }
}